Added Parser and many helper functions
The eval function calls the parser, which turns a string like this
"!(++5+--7)*9<=9+--3*4" into a string like this " ! ( 6 + 6 ) * 9 <= 9 + 2 * 4"
so that it can be easily parsed later on.
AND THEN
The eval function separates out the math parts and the boolean equation, evaluating them separately via eval and conversion functions
so a string like this " ! ( 6 + 6 ) * 9 <= 9 + 2 * 4" will cause these function calls
evalPostBool(toPostBool("! " + evalPostMath(toPostMath("( 6 + 6 ) * 9") + " <= " + evalPostMath(toPostMath("9 + 2 * 4"))
NOTE
The conversion functions toPostMath, toPostBool, evalPostMath, and evalPost bool are all called from the eval function and should be called nowhere else. Feel free to use any of the functions I wrote in the EvaluatorHelper, or add functions you would like to use.